One does not simply negotiate a change to the address architecture
The point about all participants wasn't really sharp enough for
@jariarkko, which I can see.  The thing is, even the proposed "changes
to addressing architecture would need to involve every node on the path"
doesn't really capture the magnitude of the problem.  IP packets end up
following all sorts of paths.

I'm not sure if this is clearer, but I hope it better conveys the
enormity of the class E problem.

Closes #41.