Move examples to an appendix
I've not modified these much, if at all.  I've added paragraph breaks to
some.

The only one that has changes is TCP, where I've expanded a little on
the discussion of mptcp and tfo, based on my own understanding of the
status of each.  I don't have good citations though, so I'd appreciate
help.

For #53.