We're not talking about policy; mostly
This is a hard needle to thread because policy controls are sometimes
unwanted or unexpected and that's not strictly positive.  Making a point
about exposure to unwanted or unexpected policy controls is part of the
message, but we also need to allow endpoints to make a deliberate choice
not to pass something.  I think that I managed that here, the SIP SBC
example is particularly good for that.

I'll follow up with some text on IP router alerts.