Cut the middlebox section right down
Discussion on the list has convinced me that this section is a
distraction.  I might not completely agree with Eliot when he describes
this as facile, but the point he makes in #61 is well taken: this text
is really a distraction from the central message.  It doesn't really
contribute that much.  All we really need is an acknowledgment that
middleboxes are participants too.

Closes #61.