Problems might be common, but the techniques might not be
Joel Halpern suggested that we more forcibly disclaim some of the
techniques that are mentioned in the document.

On review, the GREASE section is clear enough about limited
applicability (and also willingness), so I think we're OK there.  But
the disclaimer in the introduction could be tightened more.

Joel acknowledged that the problem exists, but that the tradeoffs
involved in applying various techniques are much more complex.  That's a
good angle on this.  I've tried to phrase that sentiment in my own
words.